treewide: code readability improvements
All downstream changes from the theme (i.e. here) will be denoted with:

/* hacks-guide change start */
and
/* hacks-guide change end */

One-line changes are marked with

/* hacks-guide change */

next to it.

Since our theme does not have linear history with upstream, this allows
quicker identification of what exactly has differentiated downstream,
as we do have a few out-of-tree patches for extra functionality.