Gemfile: pin jekyll-sass-converter to v2.2
Updating to Jekyll 4.3.0 and above causes build warnings as this
converter has switched to a different implementation (sass-embedded).
[0]

As the theme is intended to be GitHub Pages compatible (which, by the
way, has a horribly outdated Jekyll version), these warnings are not
avoidable on later Jekyll versions which have added support for the
latest version of jekyll-sass-converter.

Since Jekyll still supports the older, now-deprecated sassc-based
version [1], let's just keep it that way.

[0]: https://github.com/mmistakes/minimal-mistakes/issues/4054
[1]: https://github.com/jekyll/jekyll/blob/v4.3.2/jekyll.gemspec#L39