priiloader: mark Priiloader Wii U Forwarder as optional
There's a lot of issues getting the user to just install this one:
- User may not have Aroma installed
- It wouldn't be ideal to force the user to install Aroma for one
  application
- Making Aroma mandatory for this page wouldn't exactly be correct
  since Priiloader by itself does not require Aroma

Due to these limitations of the current Wii U environment, this one
won't be listed as a must-have. This may make it confusing and the end
result of any vWii following this guide will differ based on the user's
reading comprehension, but it doesn't make sense to remove it either,
based on provided feedback.

This may change in the future if:
- wiiu.hacks.guide fully migrates to Aroma
- More aspects of the vWii are made Aroma-only and thus is required
  for anything on this guide