Poor man's attempt at trying not to shove info into user's faces (#73)
* str2hax: make common issues inline to the step where it happens

Putting it up at the top does nothing to help the user in the event
they are actually hit with it. Move it to the relevant steps.

TODO: do we need to do it for the CTGP notice too?

* wilbrand: don't reference LetterBomb

Wilbrand is its own thing; it need not explain its differences to
LetterBomb. Just explain what it is and what it supports in a single
line.

* wilbrand: don't hint to LetterBomb in the event the installer fails to launch

While, in most cases, users will be on 4.3, this may not always be the
case. In such cases, LetterBomb will not function on the user's
console.

Just send them back to get-started and let the user pick something
sane.

* get-started: reorganize introduction points

- Move non-notices above the notices
- Don't put the "it all leads to the same thing" in a notice, it's not
  important enough to deserve such a thing

* get-started: move legacy exploit note to its own section at the bottom

This likely leads to a bigger chance of user trying to follow that
instead. The important part is in the table, so move the legacy note to
the bottom, which, if all other usual methods do not work, users will
end up falling through to.

* hbc: hide BootMii/HBC info in a dropdown

Pretty much all of this is boilerplate information that the user does
not care about (especially the boot2 stuff.)

We can hide it in a dropdown, where "Technical Details" are to be
defined.

* bluebomb: move info block above the notices

This is just going to get buried in the, quite frankly, insane amount
of notices right at the top.

TODO: How can we reduce those notices?

* cios: hide boilerplate text in dropdown

TMI.

* cios: download the IOS from NUS anyway

Given there is a whole warning that mentions skipping may have issues,
just do it from PC. It ensures a known working condition.

Remove note to skip Section I entirely.

* osc: halve the info text

side note, it sounds like an ad with it lol

* letterbomb: move info text above notices

* bluebomb: can't recover from 003 (yet....(?))

---------

Co-authored-by: thepikachugamer <44107089+Naim2000@users.noreply.github.com>