fix SIP mini link for wnd-mini.md
For an unknown reason, the developer of OSC decided to completely remove the app from the repository. This was in part of the MediaFire links being not supported anymore since MediaFire will no longer allow automatic downloads on their site, however other MediaFire apps were archived, except for this one and Korean Kii with no explanation.