Only recommend 4.3 on Wii Factory Reset (#178)
* Only recommend 4.3 on Wii Factory Reset

A helpee bricked their Wii because they saw 3.2, and ended up typing it, and since their Wii was a 4 layer Wii (way newer than RVL-CPU-20), it is unable to boot due to the simplified power circuitry. There is no reason to use an older version with this guide anyways.

* Fix small typo

forgot the b lol